Fixed lazy loading
Lazy loading can now be called in place of one-time loads.  Use new
MyObject().lazyAll() or lazyAllWhere(“condition) to receive a lazy list
as a LazyList Object.  This set can be iterated over using hasNext()
and next().  I am working on having it subscribe to utterable so it can
be used in for loops.